Implement a global mock configuration system.
Prior to this commit, mocking was done on a individual class-by-class
basis. This has two problems:

1. It isn't extensible, as BlueEyesServiceSpecification needs to know
about all mockable classes

2. It doesnt' work now that code is split across different projects.
In particular, Mongo isn't in the same project as
BlueEyesServiceSpecification

This change introduces a global mock configuraiton. The semantics are
thus:

- If a per-class key is set, use the value of that key to determine
if that class should be mocked;

- Otherwise, inspect the value of the global key

The MockConfiguration object handles manipulating the global key and
implementing the semantics above.

Upated BlueEyesServiceSpecification, ConfigurableMongo and
ConfigurableHttpClient to use this system.