Refactored async parsing to provide cleaner interface.
Per John's suggestions, JsonParser now has a parseAsync method.
Given a ParseAsyncState instance and an Option[ByteBuffer], it
will return (ParseAsyncState, Seq[Validation[Throwable, JValue]]).
The state instances are internally mutable but provide no public
methods besides copy(), or any constructors which could close over
global state.

Benchmarking suggests that for 36.7M of input data the new method is
about as efficient as the old one for chunk sizes of 10K and
greater. For chunk sizes of 1K, the new method runs in 150% of the
time, and for chunk sizes of 100 bytes, it runs in about 250% of
the time.

It might be nice to maintain some form of the old interface, since
for large streams and/or small chunk sizes, the new method does have
a lot more overhead. In many cases though it will be fine.