Silence IOException caused by stopping Mongo
We adopt the heavy handed way of catching the exception and ignoring it,
but it would be difficult to do otherwise. No matter what we check, there's
always a race window between checking and reading from the input stream.

It would have been possible to catch the IOException when reading, and then
only discarding if the reason is that the stream was closed. To do so would
require rewriting some logic that BasicIO makes available (by mistake it
turns out -- it wasn't public on SBT), so I opted for code reuse and a bit
wider net than I'd prefer.

PLATFORM-685 #finish